+++ verifier/src/main/java/com/aws/jverify/verifier/compiler/simplifications/MoveStaticMethodsToStaticType.java
@@ -120,7 +120,7 @@ public class MoveStaticMethodsToStaticType {
                         continue;
                     }
                 } else if (member instanceof JCTree.JCVariableDecl varDecl) {
-                    if (BaseDafnyGenerator.isStatic(varDecl.mods)) {
+                    if (JVerifyUtils.isStatic(varDecl.mods)) {
                         staticMembers = staticMembers.append(varDecl);
                         var varSymbol = varDecl.sym;
