+++ verifier/src/main/java/com/aws/jverify/verifier/compiler/simplifications/VerifyAnnotationCompiler.java
@@ -95,7 +95,7 @@ public class VerifyAnnotationCompiler extends TreeScanner {
 
     private void addMethodToIntervalTree(JCTree.JCMethodDecl methodDecl, boolean shouldVerify) {
         var sourceFileURI = reporter.compilationUnit.getSourceFile().toUri();
-        if (sourceFileToMethodIntervalTreeMap.containsKey(sourceFileURI) && !BaseDafnyGenerator.isSynthetic(methodDecl.getModifiers().flags)
+        if (sourceFileToMethodIntervalTreeMap.containsKey(sourceFileURI) && !JVerifyUtils.isSynthetic(methodDecl.getModifiers().flags)
                 && MethodOrLoopContractCompiler.hasImplementation(methodDecl)) {
 
             var startPos = positionCalculator.toToken(positionCalculator.getStartPos(methodDecl));
@@ -177,17 +177,17 @@ public class VerifyAnnotationCompiler extends TreeScanner {
         ShouldVerifyMode mode;
         var verifyAnnotation = annotationsByName.get(Verify.class.getName());
         if (verifyAnnotation != null) {
-            var arguments = BaseDafnyGenerator.getArguments(verifyAnnotation);
+            var arguments = JVerifyUtils.getArguments(verifyAnnotation);
             var shouldArgument = arguments.get("value");
             var should = true;
             if (shouldArgument != null) {
-                should = (boolean) BaseDafnyGenerator.getLiteralValue(shouldArgument);
+                should = (boolean) JVerifyUtils.getLiteralValue(shouldArgument);
             }
 
             var pushDownArgument = arguments.get("overrideChildren");
             var includeMembers = true;
             if (pushDownArgument != null) {
-                includeMembers = (boolean) BaseDafnyGenerator.getLiteralValue(pushDownArgument);
+                includeMembers = (boolean) JVerifyUtils.getLiteralValue(pushDownArgument);
             }
             mode = getVerifyMode(should, includeMembers);
         } else {
